refactor: copy construct verify url from components package to services package
