+++ packages/components/src/components/ConformityCredential/ConformityCredential.tsx
@@ -3,9 +3,15 @@ import _ from 'lodash';
 
 import { LoadingButton } from '@mui/lab';
 import { Table, TableBody, TableCell, TableRow, TableContainer, TableHead, Paper } from '@mui/material';
-import { generateUUID, getJsonDataFromConformityAPI, getStorageServiceLink, getValueByPath } from '@mock-app/services';
+import {
+  detectValueFromStorage,
+  generateUUID,
+  getJsonDataFromConformityAPI,
+  getStorageServiceLink,
+  getValueByPath,
+} from '@mock-app/services';
 
-import { checkStoredCredentialsConfig, detectValueFromStorage } from './utils.js';
+import { checkStoredCredentialsConfig } from './utils.js';
 import { Status, ToastMessage, toastMessage } from '../ToastMessage/ToastMessage.js';
 import {
   IConformityCredentialProps,
